+++ allauth/account/internal/flows/logout.py
@@ -7,15 +7,16 @@ from allauth.account.internal.flows.password_reset_by_code import (
 from allauth.account.internal.stagekit import clear_login
 
 
-def logout(request: HttpRequest) -> None:
+def logout(request: HttpRequest, *, show_message: bool = True) -> None:
     from allauth.account.adapter import get_adapter
     from allauth.account.views import INTERNAL_RESET_SESSION_KEY
 
     if request.user.is_authenticated:
         adapter = get_adapter()
-        adapter.add_message(
-            request, messages.SUCCESS, "account/messages/logged_out.txt"
-        )
+        if show_message:
+            adapter.add_message(
+                request, messages.SUCCESS, "account/messages/logged_out.txt"
+            )
         adapter.logout(request)
     clear_login(request)
     request.session.pop(PASSWORD_RESET_VERIFICATION_SESSION_KEY, None)