+++ allauth/idp/oidc/internal/clientkit.py
@@ -1,9 +1,11 @@
 import re
-from typing import List, Pattern
+from typing import List, Optional, Pattern, Set
 from urllib.parse import ParseResult, parse_qsl, urlparse
 
 from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
+
+from allauth.idp.oidc.models import Client
 
 
 def is_loopback(parsed_uri: ParseResult) -> bool:
@@ -120,3 +122,37 @@ def is_origin_allowed(
             return True
 
     return False
+
+
+def get_used_schemes(client: Client) -> Set[str]:
+    schemes = set()
+    for uri in client.get_redirect_uris():
+        parsed = urlparse(uri)
+        if parsed.scheme:
+            schemes.add(parsed.scheme)
+    return schemes
+
+
+def clean_post_logout_redirect_uri(
+    post_logout_redirect_uri: Optional[str], client: Optional[Client]
+) -> Optional[str]:
+    """
+    This URI SHOULD use the https scheme and MAY contain port, path, and
+    query parameter components; however, it MAY use the http scheme, provided
+    that the Client Type is confidential, as defined in Section 2.1 of OAuth 2.0
+    [RFC6749], and provided the OP allows the use of http RP URIs. The URI MAY
+    use an alternate scheme, such as one that is intended to identify a callback
+    into a native application. The value MUST have been previously registered
+    with the OP, either using the post_logout_redirect_uris Registration
+    parameter or via another mechanism. An id_token_hint is also RECOMMENDED
+    when this parameter is included.
+    """
+    allowed_schemes = {"https"}
+    if client:
+        allowed_schemes.update(get_used_schemes(client))
+        if client.type == Client.Type.CONFIDENTIAL:
+            allowed_schemes.add("http")
+    parsed = urlparse(post_logout_redirect_uri)
+    if not parsed.scheme or parsed.scheme not in allowed_schemes:
+        return None
+    return post_logout_redirect_uri