+++ allauth/idp/oidc/models.py
@@ -8,7 +8,6 @@ from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
 from allauth.idp.oidc.adapter import get_adapter
-from allauth.idp.oidc.internal.clientkit import _validate_uri_wildcard_format
 
 
 def default_client_id() -> str:
@@ -159,12 +158,16 @@ class Client(models.Model):
         return check_password(secret, self.secret)
 
     def clean_redirect_uris(self):
+        from allauth.idp.oidc.internal.clientkit import _validate_uri_wildcard_format
+
         uris = self.get_redirect_uris()
         for uri in uris:
             _validate_uri_wildcard_format(uri, self.allow_uri_wildcards)
         return uris
 
     def clean_cors_origins(self):
+        from allauth.idp.oidc.internal.clientkit import _validate_uri_wildcard_format
+
         origins = self.get_cors_origins()
         for origin in origins:
             _validate_uri_wildcard_format(origin, self.allow_uri_wildcards)